cnn_xray/requirements.txt
streamlit>=1.32.0
tensorflow==2.20.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
